+++ lib/scrollable_clean_calendar.dart
@@ -3,7 +3,7 @@ library scrollable_clean_calendar;
 import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:scrollable_clean_calendar/models/day_values_model.dart';
-import 'package:scrollable_clean_calendar/src/clean_calendar_controller.dart';
+import 'package:scrollable_clean_calendar/controllers/clean_calendar_controller.dart';
 import 'package:scrollable_clean_calendar/utils/enums.dart';
 import 'package:scrollable_clean_calendar/widgets/days_widget.dart';
 import 'package:scrollable_clean_calendar/widgets/month_widget.dart';
@@ -11,36 +11,97 @@ import 'package:scrollable_clean_calendar/widgets/weekdays_widget.dart';
 import 'package:scrollable_clean_calendar/utils/extensions.dart';
 
 class ScrollableCleanCalendar extends StatefulWidget {
+  /// The language locale
   final String locale;
+
+  /// Scroll controller
   final ScrollController? scrollController;
+
+  /// Obrigatory: The mininimum date to show
   final DateTime minDate;
+
+  /// Obrigatory: The maximum date to show
   final DateTime maxDate;
+
+  /// An initial selected date
   final DateTime? initialDateSelected;
+
+  /// The end of selected range
   final DateTime? endDateSelected;
+
+  /// If is to show or not the weekdays in calendar
   final bool showWeekdays;
+
+  /// In what weekday position the calendar is going to start
   final int weekdayStart;
+
+  /// Function when a day is tapped
   final Function(DateTime date)? onDayTapped;
+
+  /// Function when a range is selected
   final Function(DateTime minDate, DateTime? maxDate)? onRangeSelected;
+
+  /// When a date before the min date is tapped
   final Function(DateTime date)? onPreviousMinDateTapped;
+
+  /// When a date after max date is tapped
   final Function(DateTime date)? onAfterMaxDateTapped;
+
+  /// If the range is enabled
   final bool isRangeMode;
+
+  /// What layout (design) is going to be used
   final Layout? layout;
+
+  /// The space between month and calendar
   final double spaceBetweenMonthAndCalendar;
+
+  /// The space between calendars
   final double spaceBetweenCalendars;
+
+  /// The horizontal space in the calendar dates
   final double calendarCrossAxisSpacing;
+
+  /// The vertical space in the calendar dates
   final double calendarMainAxisSpacing;
+
+  /// The parent padding
   final EdgeInsetsGeometry? padding;
+
+  /// The label text style of month
   final TextStyle? monthTextStyle;
+
+  /// The label text align of month
   final TextAlign? monthTextAlign;
-  final TextStyle? weekDayTextStyle;
+
+  /// The label text align of month
+  final TextStyle? weekdayTextStyle;
+
+  /// The label text style of day
   final TextStyle? dayTextStyle;
+
+  /// The day selected background color
   final Color? daySelectedBackgroundColor;
+
+  /// The day background color
   final Color? dayBackgroundColor;
+
+  /// The day selected background color that is between day selected edges
   final Color? daySelectedBackgroundColorBetween;
+
+  /// The day disable background color
   final Color? dayDisableBackgroundColor;
-  final double radius;
+
+  /// The radius of day items
+  final double dayRadius;
+
+  /// A builder to make a customized month
   final Widget Function(BuildContext context, String month)? monthBuilder;
-  final Widget Function(BuildContext context, String weekDay)? weekDayBuilder;
+
+  /// A builder to make a customized weekday
+  final Widget Function(BuildContext context, String weekday)? weekdayBuilder;
+
+  /// A builder to make a customized day of calendar
   final Widget Function(BuildContext context, DayValues values)? dayBuilder;
 
   const ScrollableCleanCalendar({
@@ -64,20 +125,20 @@ class ScrollableCleanCalendar extends StatefulWidget {
     this.spaceBetweenMonthAndCalendar = 24,
     this.padding,
     this.monthBuilder,
-    this.weekDayBuilder,
+    this.weekdayBuilder,
     this.dayBuilder,
     this.monthTextAlign,
     this.monthTextStyle,
-    this.weekDayTextStyle,
+    this.weekdayTextStyle,
     this.daySelectedBackgroundColor,
     this.dayBackgroundColor,
     this.daySelectedBackgroundColorBetween,
     this.dayDisableBackgroundColor,
     this.dayTextStyle,
-    this.radius = 6,
+    this.dayRadius = 6,
   }) : assert(layout != null ||
             (monthBuilder != null &&
-                weekDayBuilder != null &&
+                weekdayBuilder != null &&
                 dayBuilder != null));
 
   @override
@@ -94,11 +155,12 @@ class _ScrollableCleanCalendarState extends State<ScrollableCleanCalendar> {
     List<DateTime> months = [];
 
     DateTime currentDate = DateTime(widget.minDate.year, widget.minDate.month);
+    months.add(currentDate);
 
-    while (!(currentDate.year != widget.maxDate.year &&
-        currentDate.month != widget.maxDate.month)) {
-      months.add(currentDate);
+    while (!(currentDate.year == widget.maxDate.year &&
+        currentDate.month == widget.maxDate.month)) {
       currentDate = DateTime(currentDate.year, currentDate.month + 1);
+      months.add(currentDate);
     }
 
     _cleanCalendarController = CleanCalendarController(
@@ -164,8 +226,8 @@ class _ScrollableCleanCalendarState extends State<ScrollableCleanCalendar> {
                   cleanCalendarController: _cleanCalendarController,
                   locale: widget.locale,
                   layout: widget.layout,
-                  weekDayBuilder: widget.weekDayBuilder,
-                  textStyle: widget.weekDayTextStyle,
+                  weekdayBuilder: widget.weekdayBuilder,
+                  textStyle: widget.weekdayTextStyle,
                 ),
                 AnimatedBuilder(
                   animation: _cleanCalendarController,
@@ -180,11 +242,12 @@ class _ScrollableCleanCalendarState extends State<ScrollableCleanCalendar> {
                       onAfterMaxDateTapped: widget.onAfterMaxDateTapped,
                       onPreviousMinDateTapped: widget.onPreviousMinDateTapped,
                       backgroundColor: widget.dayBackgroundColor,
-                      selectedColor: widget.daySelectedBackgroundColor,
-                      selectedColorBetween:
+                      selectedBackgroundColor:
+                          widget.daySelectedBackgroundColor,
+                      selectedBackgroundColorBetween:
                           widget.daySelectedBackgroundColorBetween,
-                      disableColor: widget.dayDisableBackgroundColor,
-                      radius: widget.radius,
+                      disableBackgroundColor: widget.dayDisableBackgroundColor,
+                      radius: widget.dayRadius,
                       textStyle: widget.dayTextStyle,
                     );
                   },